+++ src/caselawclient/models/documents/comparison.py
@@ -19,22 +19,22 @@ class Comparison(dict[str, AttributeComparison]):
         # court, date, title
         self["court"] = AttributeComparison(
             label="court",
-            this_value=this_doc.metadata["court"].value,  # type: ignore[attr-defined]
-            that_value=that_doc.metadata["court"].value,  # type: ignore[attr-defined]
-            match=this_doc.metadata["court"].value == that_doc.metadata["court"].value,  # type: ignore[attr-defined]
+            this_value=this_doc.metadata["court"].value,
+            that_value=that_doc.metadata["court"].value,
+            match=this_doc.metadata["court"].value == that_doc.metadata["court"].value,
         )
 
         self["date"] = AttributeComparison(
             label="date",
-            this_value=this_doc.metadata["date"].as_string,  # type: ignore[attr-defined]
-            that_value=that_doc.metadata["date"].as_string,  # type: ignore[attr-defined]
-            match=this_doc.metadata["date"].as_string == that_doc.metadata["date"].as_string,  # type: ignore[attr-defined]
+            this_value=this_doc.metadata["date"].as_string,
+            that_value=that_doc.metadata["date"].as_string,
+            match=this_doc.metadata["date"].as_string == that_doc.metadata["date"].as_string,
         )
         self["name"] = AttributeComparison(
             label="name",
-            this_value=this_doc.metadata["name"].value,  # type: ignore[attr-defined]
-            that_value=that_doc.metadata["name"].value,  # type: ignore[attr-defined]
-            match=this_doc.metadata["name"].value == that_doc.metadata["name"].value,  # type: ignore[attr-defined]
+            this_value=this_doc.metadata["name"].value,
+            that_value=that_doc.metadata["name"].value,
+            match=this_doc.metadata["name"].value == that_doc.metadata["name"].value,
         )
 
     def match(self) -> bool: